# Conflicts:
#	trfc-web/src/main/webapp/WEB-INF/template/businessManage/Report/ReportAllocting.jsp
#	trfc-web/src/main/webapp/WEB-INF/template/businessManage/Report/ReportManagement.jsp

+++ trfc-service/src/main/java/com/tianrui/service/bean/quality/sales/SalesBatchnum.java
@@ -102,11 +102,11 @@ public class SalesBatchnum {
     /**
      * 过磅量
      */
-    private Long weighed;
+    private Double weighed;
     /**
      * 余量
      */
-    private Long margin;
+    private Double margin;
     public String getId() {
         return id;
     }
@@ -291,19 +291,19 @@ public class SalesBatchnum {
         this.utc = utc;
     }
 
-	public Long getWeighed() {
+	public Double getWeighed() {
 		return weighed;
 	}
 
-	public void setWeighed(Long weighed) {
+	public void setWeighed(Double weighed) {
 		this.weighed = weighed;
 	}
 
-	public Long getMargin() {
+	public Double getMargin() {
 		return margin;
 	}
 
-	public void setMargin(Long margin) {
+	public void setMargin(Double margin) {
 		this.margin = margin;
 	}
